继续清理 ng-alain 示例项目中的代码；将相关示全数据改为 dotnetClub 名称
