+++ src/Discussion.Admin/ClientApp/src/app/core/startup/startup.service.ts
@@ -1,8 +1,5 @@
 import { Injectable, Injector, Inject } from '@angular/core';
-import { Router } from '@angular/router';
 import { HttpClient } from '@angular/common/http';
-import { zip } from 'rxjs';
-import { catchError } from 'rxjs/operators';
 import { MenuService, SettingsService, TitleService, ALAIN_I18N_TOKEN } from '@delon/theme';
 import { DA_SERVICE_TOKEN, ITokenService } from '@delon/auth';
 import { ACLService } from '@delon/acl';
@@ -23,61 +20,21 @@ export class StartupService {
     private injector: Injector
   ) { }
 
-  private viaHttp(resolve: any, reject: any) {
-    zip(
-      this.httpClient.get('assets/tmp/app-data.json')
-    ).pipe(
-      // 接收其他拦截器后产生的异常消息
-      catchError(([appData]) => {
-          resolve(null);
-          return [appData];
-      })
-    ).subscribe(([appData]) => {
-
-      // application data
-      const res: any = appData;
-      // 应用信息：包括站点名、描述、年份
-      this.settingService.setApp(res.app);
-      // 用户信息：包括姓名、头像、邮箱地址
-      this.settingService.setUser(res.user);
-      // ACL：设置权限为全量
-      this.aclService.setFull(true);
-      // 初始化菜单
-      this.menuService.add(res.menu);
-      // 设置页面标题的后缀
-      this.titleService.suffix = res.app.name;
-    },
-    () => { },
-    () => {
-      resolve(null);
-    });
-  }
-
-  private viaMock(resolve: any, reject: any) {
-    // const tokenData = this.tokenService.get();
-    // if (!tokenData.token) {
-    //   this.injector.get(Router).navigateByUrl('/passport/login');
-    //   resolve({});
-    //   return;
-    // }
-    // mock
+  private setupBasicInformation(resolve: any, reject: any) {
     const app: any = {
-      name: `ng-alain`,
-      description: `Ng-zorro admin panel front-end framework`
+      name: `dotnetClub`,
+      description: `.NET Core club 管理中心`
     };
     const user: any = {
       name: 'Admin',
       avatar: './assets/tmp/img/avatar.jpg',
-      email: 'cipchk@qq.com',
+      email: 'jijie.chen@someplace.com',
       token: '123456789'
     };
-    // 应用信息：包括站点名、描述、年份
     this.settingService.setApp(app);
-    // 用户信息：包括姓名、头像、邮箱地址
     this.settingService.setUser(user);
-    // ACL：设置权限为全量
     this.aclService.setFull(true);
-    // 初始化菜单
+
     this.menuService.add([
       {
         text: '主导航',
@@ -96,20 +53,15 @@ export class StartupService {
         ]
       }
     ]);
-    // 设置页面标题的后缀
+
     this.titleService.suffix = app.name;
 
     resolve({});
   }
 
   load(): Promise<any> {
-    // only works with promises
-    // https://github.com/angular/angular/issues/15088
     return new Promise((resolve, reject) => {
-      // http
-      // this.viaHttp(resolve, reject);
-      // mock：请勿在生产环境中这么使用，viaMock 单纯只是为了模拟一些数据使脚手架一开始能正常运行
-      this.viaMock(resolve, reject);
+      this.setupBasicInformation(resolve, reject);
     });
   }
 }
