--- src/Discussion.Admin/ClientApp/src/app/routes/routes-routing.module.ts
@@ -1,50 +1,0 @@
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-import { environment } from '@env/environment';
-// layout
-import { LayoutDefaultComponent } from '../layout/default/default.component';
-import { LayoutPassportComponent } from '../layout/passport/passport.component';
-// dashboard pages
-import { DashboardComponent } from './dashboard/dashboard.component';
-// passport pages
-import { UserLoginComponent } from './passport/login/login.component';
-import { UserRegisterComponent } from './passport/register/register.component';
-// single pages
-import { CallbackComponent } from './callback/callback.component';
-import { Exception403Component } from './exception/403.component';
-import { Exception404Component } from './exception/404.component';
-import { Exception500Component } from './exception/500.component';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: LayoutDefaultComponent,
-    children: [
-      { path: '', redirectTo: 'dashboard', pathMatch: 'full' },
-      { path: 'dashboard', component: DashboardComponent, data: { title: '仪表盘', titleI18n: 'dashboard' } },
-      // 业务子模块
-      // { path: 'widgets', loadChildren: './widgets/widgets.module#WidgetsModule' }
-    ]
-  },
-  // passport
-  {
-    path: 'passport',
-    component: LayoutPassportComponent,
-    children: [
-      { path: 'login', component: UserLoginComponent, data: { title: '登录', titleI18n: 'pro-login' } },
-      { path: 'register', component: UserRegisterComponent, data: { title: '注册', titleI18n: 'pro-register' } },
-    ]
-  },
-  // 单页不包裹Layout
-  { path: 'callback/:type', component: CallbackComponent },
-  { path: '403', component: Exception403Component },
-  { path: '404', component: Exception404Component },
-  { path: '500', component: Exception500Component },
-  { path: '**', redirectTo: 'dashboard' }
-];
-
-@NgModule({
-  imports: [RouterModule.forRoot(routes, { useHash: environment.useHash })],
-  exports: [RouterModule]
-})
-export class RouteRoutingModule { }
