+++ src/Discussion.Admin/ClientApp/src/app/routes/routes.module.ts
@@ -1,7 +1,11 @@
 import { NgModule } from '@angular/core';
-
 import { SharedModule } from '@shared/shared.module';
-import { RouteRoutingModule } from './routes-routing.module';
+import { Routes, RouterModule } from '@angular/router';
+import { environment } from '@env/environment';
+
+// layout
+import { LayoutDefaultComponent } from '../layout/default/default.component';
+import { LayoutPassportComponent } from '../layout/passport/passport.component';
 // dashboard pages
 import { DashboardComponent } from './dashboard/dashboard.component';
 // passport pages
@@ -14,12 +18,39 @@ import { Exception404Component } from './exception/404.component';
 import { Exception500Component } from './exception/500.component';
 
 
+
+const routes: Routes = [
+  {
+    path: '',
+    component: LayoutDefaultComponent,
+    children: [
+      { path: '', redirectTo: 'dashboard', pathMatch: 'full' },
+      { path: 'dashboard', component: DashboardComponent, data: { title: '仪表盘', titleI18n: 'dashboard' } },
+    ]
+  },
+  {
+    path: 'passport',
+    component: LayoutPassportComponent,
+    children: [
+      { path: 'login', component: UserLoginComponent, data: { title: '登录', titleI18n: 'pro-login' } },
+      { path: 'register', component: UserRegisterComponent, data: { title: '注册', titleI18n: 'pro-register' } },
+    ]
+  },
+  { path: 'callback/:type', component: CallbackComponent },
+  { path: '403', component: Exception403Component },
+  { path: '404', component: Exception404Component },
+  { path: '500', component: Exception500Component },
+  { path: '**', redirectTo: 'dashboard' }
+];
+
+
+
 const COMPONENTS = [
   DashboardComponent,
-  // passport pages
+
   UserLoginComponent,
   UserRegisterComponent,
-  // single pages
+
   CallbackComponent,
   Exception403Component,
   Exception404Component,
@@ -28,7 +59,8 @@ const COMPONENTS = [
 const COMPONENTS_NOROUNT = [];
 
 @NgModule({
-  imports: [ SharedModule, RouteRoutingModule ],
+  imports: [ SharedModule, RouterModule.forRoot(routes, { useHash: environment.useHash }) ],
+  exports: [RouterModule],
   declarations: [
     ...COMPONENTS,
     ...COMPONENTS_NOROUNT
